pyproject.toml
[project]
name = "orxhestra"
version = "0.1.4"
description = "Multi-Agent Orchestration Framework for Python"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    {name = "Nicolai M. T. Lassen"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",

    "Operating System :: OS Independent",
]
dependencies = [
    "langchain>=1.2.12",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
a2a = ["fastapi>=0.115.0"]
# LLM providers
openai = ["langchain-openai>=1.1.11"]
anthropic = ["langchain-anthropic>=0.3.0"]
google = ["langchain-google-genai>=2.1.0"]
google-vertexai = ["langchain-google-vertexai>=2.0.0"]
aws = ["langchain-aws>=0.2.0"]
azure-ai = ["langchain-azure-ai>=0.1.0"]
mistral = ["langchain-mistralai>=0.2.0"]
cohere = ["langchain-cohere>=0.3.0"]
fireworks = ["langchain-fireworks>=0.2.0"]
together = ["langchain-together>=0.2.0"]
groq = ["langchain-groq>=0.2.0"]
nvidia = ["langchain-nvidia-ai-endpoints>=0.3.0"]
huggingface = ["langchain-huggingface>=0.1.0"]
deepseek = ["langchain-deepseek>=0.1.0"]
ollama = ["langchain-ollama>=0.2.0"]
xai = ["langchain-xai>=0.2.0"]
ibm = ["langchain-ibm>=0.3.0"]
perplexity = ["langchain-perplexity>=0.1.0"]
openrouter = ["langchain-openrouter>=0.1.0"]
upstage = ["langchain-upstage>=0.3.0"]
# Infrastructure
auth = ["cryptography>=43.0", "base58>=2.1", "PyJWT>=2.8"]
database = ["sqlalchemy[asyncio]>=2.0", "aiosqlite>=0.20"]
mcp = ["fastmcp>=2.0.0"]
composer = ["pyyaml>=6.0"]
cli = ["rich>=13.0", "pyyaml>=6.0", "pyinklib>=1.1.15"]
all = [
    "fastapi>=0.115.0",
    "langchain-openai>=1.1.11",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "fastmcp>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pyinklib>=1.1.15",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "cryptography>=43.0",
    "base58>=2.1",
    "PyJWT>=2.8",
]

[project.scripts]
orx = "orxhestra.cli.app:main"

[project.urls]
Homepage = "https://github.com/NicolaiLassen/orxhestra"
Repository = "https://github.com/NicolaiLassen/orxhestra"
Issues = "https://github.com/NicolaiLassen/orxhestra/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["orxhestra*"]

[tool.setuptools.package-data]
"orxhestra.cli" = ["*.yaml"]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "build>=1.4.0",
    "twine>=6.2.0",
    "langchain-openai>=1.1.11",
    "langfuse>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
